+++ datasets/transforms.py
@@ -10,7 +10,6 @@ import torchvision.transforms as T
 import torchvision.transforms.functional as F
 
 from util.box_ops import box_xyxy_to_cxcywh
-from util.misc import interpolate
 
 
 def crop(image, target, region):
@@ -124,10 +123,6 @@ def resize(image, target, size, max_size=None):
 
     h, w = size
     target["size"] = torch.tensor([h, w])
-
-    if "masks" in target:
-        target['masks'] = interpolate(
-            target['masks'][:, None].float(), size, mode="nearest")[:, 0] > 0.5
 
     return rescaled_image, target
